Removes "executable" attribute from components
TL;DR - Components shouldn't be executable anyway.

I deal with Rack apps for my day job and use Pow for local development.

Running PHP apps with Pow installed is… difficult. In order to get this
app to boot, I had to add a trivial Rack app in order for Pow to work
with PHP.

When this Rack app serves files with an executable flag, it tries to
execute those files before sending them down the wire, as if they were
PHP files resulting in a 500 error.

Removing the executable flag prevents this from occurring.

See also: https://github.com/eric1234/rack-legacy/issues/12

This also adds an entry to .gitignore to ignore the Rack app I needed in
order to boot the app.