Add a quick and dirty MySQL to Mongo migration script
Since this is a one off script, its not particularly well designs.

Note: The MySQL store was originally used to store a small subset of the
available data returned by Twitter. The switch to Mongo stores the
entire Tweet document. Since a straight copy is not possible, the
migration script uses the Twitter REST Api to retrieve the entire
document.

Unfortunately, this means that the migration script is bound to the
Twitter Api's Rate Limiting of 180 requests every 15 minutes. That leads
to incredibly long run times.